添加结算

+++ src/main/java/com/my_ebook/entity/Car.java
@@ -62,7 +62,7 @@ public class Car {
 		this.orderMount = orderMount;
 	}
 
-	public double getTotalPrice(){
+	public float getTotalPrice(){
 		return book.getPrice() * orderMount;
 	}
